Revert to original
