+++ vibevoice/finetune/__init__.py
@@ -1,0 +1,1 @@
+# VibeVoice finetuning module
