+++ vibevoice/finetune/data_vibevoice.py
@@ -1,36 +1,39 @@
-"""
-Data processing code for VibeVoice finetuning.
-
-Adapted from https://github.com/voicepowered-ai/VibeVoice-finetuning/blob/main/src/data_vibevoice.py (MIT License)
-"""
-
 import math
-import random
-import warnings
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
 
-import librosa
 import numpy as np
 import torch
+import warnings
+import random
+
+try:
+    import librosa  # type: ignore
+except Exception:  # pragma: no cover
+    librosa = None  # Fallback: user must install librosa when using local audio paths
+
+try:
+    import resampy  # type: ignore
+except Exception:  # pragma: no cover
+    resampy = None
 
 
 def _resample_if_needed(wav: np.ndarray, orig_sr: int, target_sr: int) -> np.ndarray:
-    """
-    Check if the audio needs to be resampled, if so, resample it.
-    """
     if orig_sr == target_sr:
         return wav.astype(np.float32, copy=False)
-    return librosa.resample(
-        y=wav.astype(np.float32), orig_sr=orig_sr, target_sr=target_sr
+    if resampy is not None:
+        return resampy.resample(wav.astype(np.float32), orig_sr, target_sr)
+    if librosa is not None:
+        return librosa.resample(y=wav.astype(np.float32), orig_sr=orig_sr, target_sr=target_sr)
+    warnings.warn(
+        "No resampler available; treating audio as target_sr without resampling. Install resampy or librosa.",
+        RuntimeWarning,
     )
+    return wav.astype(np.float32, copy=False)
 
 
+# Lightweight HF-style dataset wrapper (optional). Trainer can also pass raw HF datasets directly.
 class VibeVoiceDataset:
-    """
-    HF-style dataset wrapper.
-    """
-
     def __init__(
         self,
         dataset: Any,
@@ -66,14 +69,12 @@ class VibeVoiceDataset:
             # FALLBACK: No prompt provided, so we auto-generate one from the target audio.
             try:
                 target_sr = 24000
-                wav_array = _load_audio_to_24k(
-                    item[self.audio_column], target_sr=target_sr
-                )
+                wav_array = _load_audio_to_24k(item[self.audio_column], target_sr=target_sr)
                 audio_len_seconds = len(wav_array) / target_sr
 
                 min_len_sec = min(5.0, audio_len_seconds / 4.0)
                 max_len_sec = min(15.0, audio_len_seconds / 2.0)
-
+                
                 if min_len_sec > max_len_sec:
                     min_len_sec = max_len_sec
                 max_len_sec = min(max_len_sec, audio_len_seconds)
@@ -84,31 +85,27 @@ class VibeVoiceDataset:
 
                     max_start_sample = len(wav_array) - prompt_len_samples
                     start_sample = random.randint(0, max_start_sample)
-
-                    prompt_crop = wav_array[
-                        start_sample : start_sample + prompt_len_samples
-                    ]
-
+                    
+                    prompt_crop = wav_array[start_sample : start_sample + prompt_len_samples]
+                    
                     data["voice_prompts"] = [prompt_crop]
                 else:
                     data["voice_prompts"] = None
 
             except Exception as e:
                 warnings.warn(f"Could not create voice prompt for item {idx}: {e}")
-                data["voice_prompts"] = None
+                data["voice_prompts"] = None            
         return data
 
 
-def _load_audio_to_24k(
-    audio: Union[str, np.ndarray, torch.Tensor, Dict[str, Any]],
-    *,
-    target_sr: int = 24000,
-) -> np.ndarray:
+def _load_audio_to_24k(audio: Union[str, np.ndarray, torch.Tensor, Dict[str, Any]], *, target_sr: int = 24000) -> np.ndarray:
     if isinstance(audio, np.ndarray):
         return audio.astype(np.float32)
     if isinstance(audio, torch.Tensor):
         return audio.detach().cpu().float().numpy()
     if isinstance(audio, str):
+        if librosa is None:
+            raise RuntimeError("librosa is required to load audio file paths. Please pip install librosa.")
         wav, sr = librosa.load(audio, sr=None, mono=True)
         wav = _resample_if_needed(wav, int(sr), target_sr)
         return wav
@@ -148,12 +145,8 @@ class VibeVoiceCollator:
 
         for ex in features:
             text: str = ex.get(self.text_field, "")
-            voice_prompts: Optional[List[Union[str, np.ndarray, torch.Tensor]]] = (
-                ex.get(self.voice_prompts_field)
-            )
-            target_audio: Union[str, np.ndarray, torch.Tensor, Dict[str, Any]] = ex.get(
-                self.audio_field
-            )
+            voice_prompts: Optional[List[Union[str, np.ndarray, torch.Tensor]]] = ex.get(self.voice_prompts_field)
+            target_audio: Union[str, np.ndarray, torch.Tensor, Dict[str, Any]] = ex.get(self.audio_field)
 
             # Clamp drop rate for safety
             _drop_rate = self.voice_prompt_drop_rate
@@ -164,11 +157,7 @@ class VibeVoiceCollator:
 
             proc = self.processor(
                 text=[text],
-                voice_samples=(
-                    [voice_prompts]
-                    if voice_prompts is not None and random.random() >= _drop_rate
-                    else None
-                ),
+                voice_samples=[voice_prompts] if voice_prompts is not None and random.random() >= _drop_rate else None,
                 padding=False,
                 truncation=False,
                 max_length=self.max_length,
@@ -176,14 +165,10 @@ class VibeVoiceCollator:
             )
 
             ids = proc["input_ids"][0].tolist()
-            attn = proc.get("attention_mask", torch.ones_like(proc["input_ids"]))[
-                0
-            ].tolist()
+            attn = proc.get("attention_mask", torch.ones_like(proc["input_ids"]))[0].tolist()
             speech_input_mask = proc.get("speech_input_mask")
             if speech_input_mask is None:
-                speech_input_mask = torch.zeros_like(
-                    proc["input_ids"], dtype=torch.bool
-                )
+                speech_input_mask = torch.zeros_like(proc["input_ids"], dtype=torch.bool)
             speech_input_mask_list = speech_input_mask[0].tolist()
 
             wav_target = _load_audio_to_24k(target_audio, target_sr=24000)
@@ -197,10 +182,7 @@ class VibeVoiceCollator:
                     T = None
                     try:
                         # Direct array-like with shape (T, D) or (T,)
-                        if (
-                            hasattr(enc_out, "shape")
-                            and len(getattr(enc_out, "shape", [])) >= 1
-                        ):
+                        if hasattr(enc_out, "shape") and len(getattr(enc_out, "shape", [])) >= 1:
                             T = int(enc_out.shape[0])
                         else:
                             # Nested lists/tuples or ModelOutput-like
@@ -209,10 +191,7 @@ class VibeVoiceCollator:
                             for _ in range(2):
                                 if isinstance(cand, (list, tuple)) and len(cand) > 0:
                                     cand = cand[0]
-                            if (
-                                hasattr(cand, "shape")
-                                and len(getattr(cand, "shape", [])) >= 1
-                            ):
+                            if hasattr(cand, "shape") and len(getattr(cand, "shape", [])) >= 1:
                                 T = int(cand.shape[0])
                     except Exception:
                         T = None
@@ -221,10 +200,7 @@ class VibeVoiceCollator:
             except Exception:
                 target_latent_len = None
             if target_latent_len is None:
-                target_latent_len = max(
-                    1,
-                    int(math.ceil(len(wav_target) / float(self.speech_compress_ratio))),
-                )
+                target_latent_len = max(1, int(math.ceil(len(wav_target) / float(self.speech_compress_ratio))))
 
             speech_diff_id = self.processor.tokenizer.speech_diffusion_id
             target_placeholders = [speech_diff_id] * target_latent_len
@@ -233,9 +209,7 @@ class VibeVoiceCollator:
             attn_extended = attn + [1] * target_latent_len
 
             acoustic_input_mask = speech_input_mask_list + [True] * target_latent_len
-            acoustic_loss_mask = ([False] * len(speech_input_mask_list)) + [
-                True
-            ] * target_latent_len
+            acoustic_loss_mask = ([False] * len(speech_input_mask_list)) + [True] * target_latent_len
 
             speech_end_id = self.processor.tokenizer.speech_end_id
             ids_extended.append(speech_end_id)
@@ -297,10 +271,7 @@ class VibeVoiceCollator:
                     "Tokenizer has no pad_token_id or eos_token_id; please set one or pass a valid pad id."
                 )
         for ids, attn, ain_mask, aloss_mask in zip(
-            sample_input_ids,
-            sample_attention_masks,
-            sample_acoustic_input_masks,
-            sample_acoustic_loss_masks,
+            sample_input_ids, sample_attention_masks, sample_acoustic_input_masks, sample_acoustic_loss_masks
         ):
             pad_len = max_seq_len - len(ids)
             padded_input_ids.append(ids + [pad_token_id] * pad_len)
@@ -310,28 +281,18 @@ class VibeVoiceCollator:
 
         input_ids_tensor = torch.tensor(padded_input_ids, dtype=torch.long)
         attention_mask_tensor = torch.tensor(padded_attention_masks, dtype=torch.long)
-        acoustic_input_mask_tensor = torch.tensor(
-            padded_acoustic_input_masks, dtype=torch.bool
-        )
-        acoustic_loss_mask_tensor = torch.tensor(
-            padded_acoustic_loss_masks, dtype=torch.bool
-        )
+        acoustic_input_mask_tensor = torch.tensor(padded_acoustic_input_masks, dtype=torch.bool)
+        acoustic_loss_mask_tensor = torch.tensor(padded_acoustic_loss_masks, dtype=torch.bool)
 
         if all_speech_waveforms:
             max_wave_len = max(w.shape[0] for w in all_speech_waveforms)
-            padded_speeches = np.zeros(
-                (len(all_speech_waveforms), max_wave_len), dtype=np.float32
-            )
+            padded_speeches = np.zeros((len(all_speech_waveforms), max_wave_len), dtype=np.float32)
             for i, w in enumerate(all_speech_waveforms):
                 L = w.shape[0]
                 padded_speeches[i, :L] = w
 
-            max_latent_len = (
-                max(all_speech_latent_lengths) if all_speech_latent_lengths else 1
-            )
-            speech_masks_np = np.zeros(
-                (len(all_speech_waveforms), max_latent_len), dtype=np.bool_
-            )
+            max_latent_len = max(all_speech_latent_lengths) if all_speech_latent_lengths else 1
+            speech_masks_np = np.zeros((len(all_speech_waveforms), max_latent_len), dtype=np.bool_)
             for i, L_lat in enumerate(all_speech_latent_lengths):
                 speech_masks_np[i, :L_lat] = True
 
@@ -342,16 +303,10 @@ class VibeVoiceCollator:
             for i, is_target in enumerate(per_segment_is_target):
                 if is_target:
                     speeches_loss_input_np[i] = speech_masks_np[i]
-            speeches_loss_input_tensor = torch.tensor(
-                speeches_loss_input_np, dtype=torch.bool
-            )
+            speeches_loss_input_tensor = torch.tensor(speeches_loss_input_np, dtype=torch.bool)
 
             # Semantic features
-            if (
-                self.compute_semantics
-                and hasattr(self.processor, "semantic_tokenizer")
-                and self.processor.semantic_tokenizer is not None
-            ):
+            if self.compute_semantics and hasattr(self.processor, "semantic_tokenizer") and self.processor.semantic_tokenizer is not None:
                 sem_feats: List[np.ndarray] = []
                 for w in all_speech_waveforms:
                     try:
@@ -361,31 +316,22 @@ class VibeVoiceCollator:
                     except Exception:
                         sem = np.zeros((0, self.semantic_vae_dim), dtype=np.float32)
                     if sem.ndim != 2:
-                        raise RuntimeError(
-                            f"Semantic tokenizer returned unexpected shape {sem.shape}. Expect [T, D]."
-                        )
+                        raise RuntimeError(f"Semantic tokenizer returned unexpected shape {sem.shape}. Expect [T, D].")
                     L = sem.shape[0]
                     D = sem.shape[1]
                     if D != self.semantic_vae_dim:
                         if D < self.semantic_vae_dim:
-                            pad_d = np.zeros(
-                                (L, self.semantic_vae_dim - D), dtype=np.float32
-                            )
+                            pad_d = np.zeros((L, self.semantic_vae_dim - D), dtype=np.float32)
                             sem = np.concatenate([sem, pad_d], axis=1)
                         else:
                             sem = sem[:, : self.semantic_vae_dim]
                     if L < max_latent_len:
-                        pad = np.zeros(
-                            (max_latent_len - L, self.semantic_vae_dim),
-                            dtype=np.float32,
-                        )
+                        pad = np.zeros((max_latent_len - L, self.semantic_vae_dim), dtype=np.float32)
                         sem = np.concatenate([sem, pad], axis=0)
                     elif L > max_latent_len:
                         sem = sem[:max_latent_len]
                     sem_feats.append(sem.astype(np.float32))
-                speech_semantic_tensors = torch.tensor(
-                    np.stack(sem_feats, axis=0), dtype=torch.float32
-                )
+                speech_semantic_tensors = torch.tensor(np.stack(sem_feats, axis=0), dtype=torch.float32)
             else:
                 # Semantic tokenizer unavailable while semantics are required for training.
                 # Raise to avoid silently degrading alignment with zeroed features.
@@ -402,9 +348,7 @@ class VibeVoiceCollator:
         if self.debug_checks:
             assert (input_ids_tensor >= 0).all(), "input_ids contains negative indices"
             if speech_tensors_tensor is not None:
-                assert (
-                    speech_tensors_tensor.dim() == 2
-                ), "Expected speech_tensors 2D [segments, samples]"
+                assert speech_tensors_tensor.dim() == 2, "Expected speech_tensors 2D [segments, samples]"
 
         return {
             "input_ids": input_ids_tensor,
